# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Baba Abdullah Kitchen API
  description: |
    Microservices REST API for Baba Abdullah Kitchen PWA dashboard.
    Supports product management, inventory, orders, sales reports, and promotions.
  version: 1.0.0
  contact:
    name: Baba Abdullah Kitchen
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://baba-menu.umsoftware.id
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SalesReport:
      type: object
      properties:
        period_label: { type: string }
        order_count: { type: integer }
        total_revenue: { type: number }
    ProfitReport:
      type: object
      properties:
        period_label: { type: string }
        order_count: { type: integer }
        total_revenue: { type: number }
        total_profit: { type: number }
    TopVariant:
      type: object
      properties:
        variant_id: { type: integer }
        product_name: { type: string }
        category_name: { type: string }
        variant_label: { type: string }
        pcs: { type: string }
        total_sold: { type: integer }
        total_revenue: { type: number }
...
paths:
...
  /api/reports/sales:
    get:
      summary: Sales totals by period
      tags: [Reports]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: period
          schema: { type: string, enum: [daily, weekly, monthly], default: daily }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200':
          description: Sales report data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SalesReport'
  /api/reports/profit:
    get:
      summary: Profit totals by period
      tags: [Reports]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: period
          schema: { type: string, enum: [daily, weekly, monthly], default: daily }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200':
          description: Profit report data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProfitReport'
  /api/reports/top-variants:
    get:
      summary: Best selling variants
      tags: [Reports]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: Top variants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TopVariant'
  /api/reports/sales/table:
    get:
      summary: Tabular sales breakdown
      tags: [Reports]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200':
          description: Tabular data
  /api/reports/export/excel:
    get:
      summary: Export sales as Excel file
      tags: [Reports]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200':
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
...